+++ Tests/Functional/Controller/Fixtures/XclassFrontendUser.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Onetimeaccount\Tests\Functional\Controller\Fixtures;
+
+use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser;
+
+final class XclassFrontendUser extends FrontendUser
+{
+}